#1 데이터 및 그래프 반영

+++ app/src/main/java/com/hipaduck/timerweb/data/TimerWebRepository.kt
@@ -5,7 +5,7 @@ interface TimerWebRepository {
 
     suspend fun getValidDates(): List<String>
 
-    suspend fun putValueFromDateKey(dateKey: String, sec: Int)
+    suspend fun putValueFromDateKey(dateKey: String, sec: Long)
 
-    suspend fun getValueFromDateKey(dateKey: String): Int
+    suspend fun getValueFromDateKey(dateKey: String): Long
 }
